requirements.txt
Flask==3.0.3
Pillow==10.4.0
numpy==2.0.1
matplotlib==3.9.1
gunicorn==21.2.0